@@ -1,19 +1,19 @@
 /**
 *  SwiftUIMasonry
-*  Copyright (c) Ciaran O'Brien 2022
+*  Copyright (c) Ciaran O'Brien 2025
 *  MIT license, see LICENSE file for details
 */
 
 import SwiftUI
 
 /// Constants that define the number of lines in a masonry view.
-public enum MasonryLines {
+public enum MasonryLines: ExpressibleByIntegerLiteral {
     
     /// A variable number of lines.
     ///
-    /// This case uses the provided `sizeConstraint` to decide the exact
+    /// This case uses the provided `lengthConstraint` to decide the exact
     /// number of lines.
-    case adaptive(sizeConstraint: AdaptiveSizeConstraint)
+    case adaptive(lengthConstraint: AdaptiveLengthConstraint)
     
     /// A constant number of lines.
     case fixed(Int)
@@ -24,18 +24,18 @@ public extension MasonryLines {
     
     /// A variable number of lines.
     ///
-    /// This case uses the provided `minSize` to decide the exact number of
+    /// This case uses the provided `minLength` to decide the exact number of
     /// lines.
-    static func adaptive(minSize: CGFloat) -> MasonryLines {
-        .adaptive(sizeConstraint: .min(minSize))
+    static func adaptive(minLength: CGFloat) -> MasonryLines {
+        .adaptive(lengthConstraint: .min(minLength))
     }
     
     /// A variable number of lines.
     ///
-    /// This case uses the provided `maxSize` to decide the exact number of
+    /// This case uses the provided `maxLength` to decide the exact number of
     /// lines.
-    static func adaptive(maxSize: CGFloat) -> MasonryLines {
-        .adaptive(sizeConstraint: .max(maxSize))
+    static func adaptive(maxLength: CGFloat) -> MasonryLines {
+        .adaptive(lengthConstraint: .max(maxLength))
     }
 }
 
@@ -43,12 +43,19 @@ public extension MasonryLines {
 public extension MasonryLines {
     /// Constants that constrain the bounds of an adaptive line in a masonry
     /// view.
-    enum AdaptiveSizeConstraint: Equatable {
+    enum AdaptiveLengthConstraint: Equatable {
         
-        /// The minimum size of a line in a given axis.
+        /// The minimum length of a line in a given axis.
         case min(CGFloat)
         
-        /// The maximum size of a line in a given axis.
+        /// The maximum length of a line in a given axis.
         case max(CGFloat)
+    }
+}
+
+
+public extension MasonryLines {
+    init(integerLiteral value: Int) {
+        self = .fixed(value)
     }
 }
